move x86 bitutils.h implementations to util/arch/x86/bitutils.h
